Also make packet handler common for DirectRouter impls
