+++ crates/ergot/src/interface_manager/profiles/direct_router/tokio_serial_5.rs
@@ -3,10 +3,10 @@
 //! This implementation can be used to connect to a number of direct edge serial devices.
 
 use crate::{
-    Header,
     interface_manager::{
         InterfaceState, Profile,
         interface_impls::tokio_serial_cobs::TokioSerialInterface,
+        profiles::direct_router::{DirectRouter, process_frame},
         utils::{
             cobs_stream::Sink,
             std::{
@@ -17,7 +17,6 @@ use crate::{
         },
     },
     net_stack::NetStackHandle,
-    wire_frames::de_frame,
 };
 use bbq2::{prod_cons::stream::StreamConsumer, traits::bbqhdl::BbqHandle};
 use cobs::max_encoding_overhead;
@@ -29,8 +28,6 @@ use tokio::{
     select,
 };
 use tokio_serial_v5::{SerialPortBuilderExt, SerialStream};
-
-use super::DirectRouter;
 
 #[derive(Debug, PartialEq)]
 pub enum Error {
@@ -149,47 +146,7 @@ where
                     FeedResult::DecodeError(new_wind) => new_wind,
                     FeedResult::Success { data, remaining }
                     | FeedResult::SuccessInput { data, remaining } => {
-                        // Successfully de-cobs'd a packet, now we need to
-                        // do something with it.
-                        if let Some(mut frame) = de_frame(data) {
-                            // If the message comes in and has a src net_id of zero,
-                            // we should rewrite it so it isn't later understood as a
-                            // local packet.
-                            if frame.hdr.src.network_id == 0 {
-                                assert_ne!(
-                                    frame.hdr.src.node_id, 0,
-                                    "we got a local packet remotely?"
-                                );
-                                assert_ne!(
-                                    frame.hdr.src.node_id, 1,
-                                    "someone is pretending to be us?"
-                                );
-
-                                frame.hdr.src.network_id = self.net_id;
-                            }
-                            // TODO: if the destination IS self.net_id, we could rewrite the
-                            // dest net_id as zero to avoid a pass through the interface manager.
-                            //
-                            // If the dest is 0, should we rewrite the dest as self.net_id? This
-                            // is the opposite as above, but I dunno how that will work with responses
-                            let hdr = frame.hdr.clone();
-                            let hdr: Header = hdr.into();
-
-                            let res = match frame.body {
-                                Ok(body) => self.nsh.stack().send_raw(&hdr, frame.hdr_raw, body),
-                                Err(e) => self.nsh.stack().send_err(&hdr, e),
-                            };
-                            match res {
-                                Ok(()) => {}
-                                Err(e) => {
-                                    // TODO: match on error, potentially try to send NAK?
-                                    warn!("recv->send error: {e:?}");
-                                }
-                            }
-                        } else {
-                            warn!("Decode error! Ignoring frame on net_id {}", self.net_id);
-                        }
-
+                        process_frame(self.net_id, data, &self.nsh);
                         remaining
                     }
                 };
